test(pci-users): move test files to the same location of files
